requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
google-generativeai
langchain_community
docx
python-dotenv